+++ src/main/java/com/example/tripKo/domain/member/api/MemberController.java
@@ -4,10 +4,17 @@ import com.example.tripKo._core.utils.ApiUtils;
 import com.example.tripKo.domain.member.application.MemberService;
 import com.example.tripKo.domain.member.dto.response.RestaurantReservationResponse;
 import java.util.List;
+
+import com.example.tripKo.domain.place.dto.request.RestaurantReservationConfirmRequest;
+import com.example.tripKo.domain.place.dto.request.RestaurantReservationSelectRequest;
+import com.example.tripKo.domain.place.dto.response.info.RestaurantReservationConfirmResponse;
+import com.example.tripKo.domain.place.dto.response.info.RestaurantReservationSelectResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @Validated
@@ -15,12 +22,28 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class MemberController {
 
+
   private final MemberService memberService;
+
 
   @GetMapping("/userinfo/reservations/restaurant")
   public ResponseEntity<?> getRestaurantReservationInfo() {
     List<RestaurantReservationResponse> response = memberService.getRestaurantReservationInfo();
     ApiUtils.ApiResult<?> apiResult = ApiUtils.success(response);
+    return ResponseEntity.ok(apiResult);
+  }
+
+  @PostMapping("/restaurant/bookings/calendar")
+  public ResponseEntity<?> selectReservationDate(@RequestBody @Validated RestaurantReservationSelectRequest requestDTOs) {
+    RestaurantReservationSelectResponse responseDTO = memberService.selectRestaurantReservationDate(requestDTOs);
+    ApiUtils.ApiResult<?> apiResult = ApiUtils.success(responseDTO);
+    return ResponseEntity.ok(apiResult);
+  }
+
+  @PostMapping("/restaurant/bookings")
+  public ResponseEntity<?> confirmReservation(@RequestBody @Validated RestaurantReservationConfirmRequest requestDTOs) {
+    RestaurantReservationConfirmResponse responseDTO = memberService.confirmRestaurantReservation(requestDTOs);
+    ApiUtils.ApiResult<?> apiResult = ApiUtils.success(responseDTO);
     return ResponseEntity.ok(apiResult);
   }
